Fix miscalculated width
When slick is redrawing with different options, as we can set in the responsive settings,
than slick sets a wrong slide width and the slides are not centered.
On webkit browsers, it sometimes happens that at first pageload, that slick also
calculates a wrong width and slides are not centered until the first slideaction.